Struct labwhere::models::location::Location

pub struct Location {
    pub id: u32,
    pub name: String,
    pub barcode: Option<String>,
    location_type_id: u32,
}

Location of the Labware

Fields
id: u32

ID of the location record

name: String

Name of the location

barcode: Option<String>

The barcode of the location

location_type_id: u32

The id of the location_type

Implementations

impl<'a> Location

Implementation of the Location struct

fn new( id: u32, name: String, location_type_id: u32, barcode: Option<String>, ) -> Result<Location, NameFormatError>

Create a new Location

Examples

use location::Location;
let location_result_1 = Location::new(1, "Building".to_string()); // Returns a Result
let location = location_result_1.unwrap();    // Call unwrap or use match with Err and Ok branches.

let location_result_2 = Location::new(1, "Building".to_string()); // Returns a Result
let location = match location_result_2 {
    // References on Panic or not to panic?
    // https://doc.rust-lang.org/beta/book/ch09-03-to-panic-or-not-to-panic.html
    Err(e) => panic!("{:?}", e),
    Ok(result) => result
};

pub(crate) async fn create( name: String, location_type_id: u32, connection: &mut SqliteConnection, ) -> Result<Location, Error>

Create a new Location

Examples

use location::Location;
let location = Location::create("location1".to_string(), 1).await.unwrap();

pub(crate) async fn find_by_barcode( barcode: String, connection: &mut SqliteConnection, ) -> Result<Location, NotFoundError>

Find a location by barcode

Examples

use location::Location;
let mut connection = init_db("sqlite::memory:").await.unwrap();
let location = Location::find_by_barode("lw-location1-1".to_string(), &mut connection).await.unwrap();

pub fn unknown() -> &'a Location

Create a new unknown location

Examples

use location::Location;
let location = Location::unknown();

fn create_barcode(&mut self) -> String

Creates a barcode Barcode format: lw-{name trimmed and spaces replaced with "-"}-{id}

fn validate_name(name: String) -> bool

Validate the name of the location for a certain format Validations: 1. Name must be between 1 and 60 characters 2. Name must only contain alphanumeric characters, hyphens, spaces, and parentheses